requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.2
django-cleanup==9.0.0
django-cors-headers==4.5.0
django-crispy-forms==2.3
django-filter==24.3
django-phonenumber-field==8.0.0
django-resized==1.0.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.1
phonenumbers==8.13.47
pillow==10.4.0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
uritemplate==4.1.1
